+++ src/main/java/com/techeer/fashioncloud/domain/geo/constant/GeoConstant.java
@@ -1,6 +1,5 @@
-package com.techeer.fashioncloud.domain.weather.constant;
+package com.techeer.fashioncloud.domain.geo.constant;
 
 public class GeoConstant {
-
     public static final String GEO_BASE_URL = "https://dapi.kakao.com/v2/local/geo/coord2regioncode.json";
 }
